Lib: Force output type in ip4_addr constructor
Fixes type issue when u64 is pushed into it.

+++ lib/ip.h
@@ -67,7 +67,7 @@ typedef struct ip4_addr {
 
 typedef u32 ip4_addr;
 
-#define _MI4(x) (x)
+#define _MI4(x) ((u32) (x))
 #define _I(x) (x)
 
 #endif
